fix(decompose): decompose sections into components to increase readability
