+++ apps/web/app/page.tsx
@@ -1,904 +1,32 @@
-"use client";
-
-import Brand from "@/components/Brand";
-import {
-  Avatar,
-  AvatarFallback,
-  AvatarImage,
-} from "@workspace/ui/components/avatar";
-import { Button } from "@workspace/ui/components/button";
-import { Input } from "@workspace/ui/components/input";
-import { IndianRupee, Search } from "lucide-react";
-import type React from "react";
-
-import { useState } from "react";
+import Cta from "@/components/home/cta";
+import FeaturedProperties from "@/components/home/featured-properties";
+import HowItWorks from "@/components/home/how-it-works";
+import MainHero from "@/components/home/main-hero";
+import PopularCities from "@/components/home/popular-cities";
+import Testimonials from "@/components/home/testimonials";
+import WhyChooseUs from "@/components/home/why-choose-us";
 
 export default function HomePage() {
-  const [mobileMenuOpen, setMobileMenuOpen] = useState(false);
-  const [searchLocation, setSearchLocation] = useState("");
-  const [budget, setBudget] = useState("");
-
-  const handleLocationChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setSearchLocation(e.target.value);
-  };
-
-  const handleBudgetChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setBudget(e.target.value);
-  };
-
-  const featuredProperties = [
-    {
-      id: 1,
-      name: "Elite Residency",
-      location: "Koramangala, Bangalore",
-      price: "₹12,000",
-      rating: 4.8,
-      reviews: 124,
-      image: "/pg-images/elite.png?height=200&width=300",
-      amenities: ["WiFi", "AC", "Parking", "Food"],
-      type: "Single Room",
-      verified: true,
-      liked: false,
-    },
-    {
-      id: 2,
-      name: "Urban Nest",
-      location: "Gurgaon, Delhi NCR",
-      price: "₹15,000",
-      rating: 4.6,
-      reviews: 89,
-      image: "/pg-images/urban.png?height=200&width=300",
-      amenities: ["WiFi", "Gym", "Food", "Laundry"],
-      type: "Shared Room",
-      verified: true,
-      liked: true,
-    },
-    {
-      id: 3,
-      name: "Comfort Zone",
-      location: "Bandra, Mumbai",
-      price: "₹18,000",
-      rating: 4.9,
-      reviews: 156,
-      image: "/pg-images/comfort.png",
-      amenities: ["WiFi", "AC", "Food", "Security"],
-      type: "Single Room",
-      verified: true,
-      liked: false,
-    },
-    {
-      id: 4,
-      name: "Student Hub",
-      location: "Anna Nagar, Chennai",
-      price: "₹10,000",
-      rating: 4.5,
-      reviews: 67,
-      image: "/pg-images/sikkim.png",
-      amenities: ["WiFi", "Study Room", "Food", "AC"],
-      type: "Shared Room",
-      verified: true,
-      liked: false,
-    },
-  ];
-
-  const cities = [
-    {
-      name: "Bangalore",
-      properties: 150,
-      image: "/bangalore.png?height=120&width=120",
-    },
-    {
-      name: "Mumbai",
-      properties: 200,
-      image: "/mumbai.png?height=120&width=120",
-    },
-    {
-      name: "Delhi NCR",
-      properties: 180,
-      image: "/delhi.png?height=120&width=120",
-    },
-    {
-      name: "Pune",
-      properties: 120,
-      image: "/pune.png?height=120&width=120",
-    },
-    {
-      name: "Chennai",
-      properties: 90,
-      image: "/chennai.png?height=120&width=120",
-    },
-    {
-      name: "Hyderabad",
-      properties: 110,
-      image: "/hyderabad.png?height=120&width=120",
-    },
-  ];
-
-  const testimonials = [
-    {
-      name: "Ananya Sharma",
-      location: "Bangalore",
-      rating: 5,
-      text: "Found the perfect PG within my budget. The booking process was super smooth and the property exactly matched the photos!",
-      avatar: "/placeholder.svg?height=50&width=50",
-    },
-    {
-      name: "Rohit Kumar",
-      location: "Mumbai",
-      rating: 5,
-      text: "Great platform with verified properties. Saved me weeks of searching. Highly recommend for anyone looking for PG accommodation.",
-      avatar: "/placeholder.svg?height=50&width=50",
-    },
-    {
-      name: "Priya Patel",
-      location: "Delhi",
-      rating: 4,
-      text: "Excellent service and genuine listings. The support team was very helpful throughout the booking process.",
-      avatar: "/placeholder.svg?height=50&width=50",
-    },
-  ];
-
   return (
     <div className="min-h-screen bg-white">
-      {/* Header */}
-      <header className="bg-white shadow-sm border-b sticky top-0 z-50">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-          <div className="flex justify-between items-center h-16">
-            <div className="flex items-center space-x-2">
-              <div className="w-8 h-8  rounded-lg flex items-center justify-center">
-                <Brand />
-              </div>
-              <p className="text-primary font-bold text-4xl">Baasthan</p>
-            </div>
-
-            {/* Desktop Navigation */}
-            <nav className="hidden md:flex items-center space-x-8">
-              <a
-                href="#"
-                className="text-gray-700 hover:text-primary font-medium transition-colors"
-              >
-                Properties
-              </a>
-              <button className="border border-primary text-primary hover:bg-primary/10 bg-transparent px-4 py-2 rounded-md font-medium transition-colors">
-                List Property
-              </button>
-            </nav>
-
-            {/* Mobile menu button */}
-            <div className="md:hidden">
-              <button
-                onClick={() => setMobileMenuOpen(!mobileMenuOpen)}
-                className="p-2 rounded-md text-gray-400 hover:text-gray-500 hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-muted-foreground"
-              >
-                {mobileMenuOpen ? (
-                  <svg
-                    className="w-6 h-6"
-                    fill="none"
-                    stroke="currentColor"
-                    viewBox="0 0 24 24"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth={2}
-                      d="M6 18L18 6M6 6l12 12"
-                    />
-                  </svg>
-                ) : (
-                  <svg
-                    className="w-6 h-6"
-                    fill="none"
-                    stroke="currentColor"
-                    viewBox="0 0 24 24"
-                  >
-                    <path
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                      strokeWidth={2}
-                      d="M4 6h16M4 12h16M4 18h16"
-                    />
-                  </svg>
-                )}
-              </button>
-            </div>
-          </div>
-
-          {/* Mobile Navigation */}
-          {mobileMenuOpen && (
-            <div className="md:hidden py-4 border-t">
-              <div className="flex flex-col space-y-4">
-                <a
-                  href="#"
-                  className="text-gray-700 hover:text-primary font-medium"
-                >
-                  Properties
-                </a>
-                <div className="pt-2">
-                  <button className="w-full border  border-primary text-primary bg-transparent px-4 py-2 rounded-md font-medium transition-colors">
-                    List Property
-                  </button>
-                </div>
-              </div>
-            </div>
-          )}
-        </div>
-      </header>
-
       {/* Hero Section */}
-      <section className="bg-gradient-to-br from-blue-50 to-white py-16 lg:py-24">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-          <div className="text-center mb-12">
-            <h1 className="text-4xl lg:text-6xl font-bold text-gray-900 mb-6">
-              Find Your Perfect{" "}
-              <span className="text-primary">PG Accommodation</span>
-            </h1>
-            <p className="text-xl text-gray-600 max-w-3xl mx-auto mb-8">
-              Discover verified PG accommodations across India with transparent
-              pricing, genuine reviews, and instant booking.
-            </p>
-
-            {/* Search Bar */}
-            <div className="max-w-4xl mx-auto">
-              <div className="bg-white p-6 shadow-lg border-0 rounded-lg">
-                <div className="grid md:grid-cols-4 gap-4">
-                  <div className="md:col-span-2">
-                    <div className="relative">
-                      {/* <svg
-                        className="absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-400 w-5 h-5"
-                        fill="currentColor"
-                        viewBox="0 0 20 20"
-                      >
-                        <path
-                          fillRule="evenodd"
-                          d="M5.05 4.05a7 7 0 119.9 9.9L10 18.9l-4.95-4.95a7 7 0 010-9.9zM10 11a2 2 0 100-4 2 2 0 000 4z"
-                          clipRule="evenodd"
-                        />
-                      </svg> */}
-                      {/* <input
-                        type="text"
-                        placeholder="Enter city or locality"
-                        value={searchLocation}
-                        onChange={handleLocationChange}
-                        className="pl-10 h-12 w-full border border-gray-200 rounded-md  focus:border-primary focus:outline-none px-3"
-                      /> */}
-
-                      <Input
-                        type="text"
-                        icon={Search}
-                        placeholder="Enter city or locality"
-                        value={searchLocation}
-                        onChange={handleLocationChange}
-                      />
-                    </div>
-                  </div>
-                  <div>
-                    {/* <input
-                        type="text"
-                        placeholder="Budget (₹)"
-                        value={budget}
-                        onChange={handleBudgetChange}
-                        className="h-12 w-full border border-gray-200 rounded-md focus:border-blue-500 focus:outline-none px-3 "
-                      /> */}
-                    {/* <Slider defaultValue={[33]} max={100} step={1} /> */}
-
-                    <Input
-                      type="text"
-                      icon={IndianRupee}
-                      placeholder="Budget"
-                      value={budget}
-                      onChange={handleBudgetChange}
-                    />
-                  </div>
-                  <Button>
-                    <Search />
-                    Search
-                  </Button>
-                </div>
-              </div>
-            </div>
-
-            {/* Quick Stats */}
-            <div className="grid grid-cols-2 md:grid-cols-4 gap-8 mt-16">
-              <div className="text-center">
-                <div className="text-3xl font-bold text-primary mb-2">
-                  1000+
-                </div>
-                <div className="text-gray-600">Verified Properties</div>
-              </div>
-              <div className="text-center">
-                <div className="text-3xl font-bold text-primary mb-2">50+</div>
-                <div className="text-gray-600">Cities Covered</div>
-              </div>
-              <div className="text-center">
-                <div className="text-3xl font-bold text-primary mb-2">25K+</div>
-                <div className="text-gray-600">Happy Residents</div>
-              </div>
-              <div className="text-center">
-                <div className="text-3xl font-bold text-primary mb-2">4.8★</div>
-                <div className="text-gray-600">Average Rating</div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </section>
-
+      <MainHero />
       {/* Popular Cities */}
-      <section className="py-16 bg-white">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-          <div className="text-center mb-12">
-            <h2 className="text-3xl font-bold text-gray-900 mb-4">
-              Popular Cities
-            </h2>
-            <p className="text-gray-600 text-lg">
-              Explore PG accommodations in top cities across India
-            </p>
-          </div>
-
-          <div className="grid grid-cols-2 md:grid-cols-3 lg:grid-cols-6 gap-6">
-            {cities.map((city, index) => (
-              <div
-                key={index}
-                className="bg-white rounded-lg shadow-md border-0 hover:shadow-lg transition-all duration-300 hover:-translate-y-1 cursor-pointer group"
-              >
-                <div className="p-6 text-center">
-                  <div className="w-16 h-16  rounded-full flex items-center justify-center mx-auto mb-4   transition-colors overflow-hidden">
-                    <img
-                      src={city.image || "/placeholder.svg"}
-                      alt={city.name}
-                      className="w-10 h-10 rounded-full object-cover"
-                    />
-                    {/* <Bangalore height={32} width={32} scale={1.5} /> */}
-                    <a
-                      href="https://www.flaticon.com/free-icons/bangalore"
-                      title="bangalore icons"
-                      className="hidden"
-                    >
-                      Bangalore icons created by Daniel ceha - Flaticon
-                    </a>
-                    <a
-                      href="https://www.flaticon.com/free-icons/india-gate"
-                      title="india gate icons"
-                      hidden
-                    >
-                      India gate icons created by Icon home - Flaticon
-                    </a>
-                  </div>
-                  <h3 className="font-semibold text-gray-900 mb-1">
-                    {city.name}
-                  </h3>
-                  <p className="text-sm text-gray-500">
-                    {city.properties} Properties
-                  </p>
-                </div>
-              </div>
-            ))}
-          </div>
-        </div>
-      </section>
-
+      <PopularCities />
       {/* Featured Properties */}
-      <section className="py-16 bg-gray-50">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-          <div className="flex justify-between items-center mb-12">
-            <div>
-              <h2 className="text-3xl font-bold text-gray-900 mb-4">
-                Featured Properties
-              </h2>
-              <p className="text-gray-600 text-lg">
-                Handpicked premium PG accommodations
-              </p>
-            </div>
-            <button className="border border-primary text-primary hover:bg-primary/10 bg-transparent px-4 py-2 rounded-md font-medium transition-colors flex items-center">
-              View All
-              <svg
-                className="w-4 h-4 ml-2"
-                fill="none"
-                stroke="currentColor"
-                viewBox="0 0 24 24"
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth={2}
-                  d="M9 5l7 7-7 7"
-                />
-              </svg>
-            </button>
-          </div>
-
-          <div className="grid md:grid-cols-2 lg:grid-cols-4 gap-6">
-            {featuredProperties.map((property) => (
-              <div
-                key={property.id}
-                className="bg-white rounded-lg shadow-md border-0 hover:shadow-xl transition-all duration-300 overflow-hidden cursor-pointer group"
-              >
-                <div className="relative">
-                  <img
-                    src={property.image || "/placeholder.svg"}
-                    alt={property.name}
-                    className="w-full h-48 object-cover group-hover:scale-105 transition-transform duration-300"
-                  />
-                  <div className="absolute top-3 left-3">
-                    {property.verified && (
-                      <span className="bg-green-500 text-white px-2 py-1 rounded-md text-xs font-medium flex items-center">
-                        <svg
-                          className="w-3 h-3 mr-1"
-                          fill="currentColor"
-                          viewBox="0 0 20 20"
-                        >
-                          <path
-                            fillRule="evenodd"
-                            d="M10 18a8 8 0 100-16 8 8 0 000 16zm3.707-9.293a1 1 0 00-1.414-1.414L9 10.586 7.707 9.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l4-4z"
-                            clipRule="evenodd"
-                          />
-                        </svg>
-                        Verified
-                      </span>
-                    )}
-                  </div>
-                  <div className="absolute top-3 right-3 flex space-x-2">
-                    <button
-                      className={`w-8 h-8 p-0 rounded-full bg-white/80 hover:bg-white transition-colors ${
-                        property.liked ? "text-red-500" : "text-gray-600"
-                      }`}
-                    >
-                      <svg
-                        className="w-4 h-4 mx-auto"
-                        fill={property.liked ? "currentColor" : "none"}
-                        stroke="currentColor"
-                        viewBox="0 0 24 24"
-                      >
-                        <path
-                          strokeLinecap="round"
-                          strokeLinejoin="round"
-                          strokeWidth={2}
-                          d="M4.318 6.318a4.5 4.5 0 000 6.364L12 20.364l7.682-7.682a4.5 4.5 0 00-6.364-6.364L12 7.636l-1.318-1.318a4.5 4.5 0 00-6.364 0z"
-                        />
-                      </svg>
-                    </button>
-                    <button className="w-8 h-8 p-0 rounded-full bg-white/80 hover:bg-white text-gray-600 transition-colors">
-                      <svg
-                        className="w-4 h-4 mx-auto"
-                        fill="none"
-                        stroke="currentColor"
-                        viewBox="0 0 24 24"
-                      >
-                        <path
-                          strokeLinecap="round"
-                          strokeLinejoin="round"
-                          strokeWidth={2}
-                          d="M8.684 13.342C8.886 12.938 9 12.482 9 12c0-.482-.114-.938-.316-1.342m0 2.684a3 3 0 110-2.684m0 2.684l6.632 3.316m-6.632-6l6.632-3.316m0 0a3 3 0 105.367-2.684 3 3 0 00-5.367 2.684zm0 9.316a3 3 0 105.367 2.684 3 3 0 00-5.367-2.684z"
-                        />
-                      </svg>
-                    </button>
-                  </div>
-                  <div className="absolute bottom-3 left-3">
-                    <span className="bg-white/90 text-gray-900 px-2 py-1 rounded-md text-xs font-medium">
-                      {property.type}
-                    </span>
-                  </div>
-                </div>
-
-                <div className="p-4">
-                  <div className="flex justify-between items-start mb-2">
-                    <h3 className="font-semibold text-gray-900 group-hover:text-primary transition-colors">
-                      {property.name}
-                    </h3>
-                    <div className="text-right">
-                      <div className="text-lg font-bold text-primary">
-                        {property.price}
-                      </div>
-                      <div className="text-xs text-gray-500">per month</div>
-                    </div>
-                  </div>
-
-                  <div className="flex items-center text-gray-600 mb-3">
-                    <svg
-                      className="w-4 h-4 mr-1"
-                      fill="currentColor"
-                      viewBox="0 0 20 20"
-                    >
-                      <path
-                        fillRule="evenodd"
-                        d="M5.05 4.05a7 7 0 119.9 9.9L10 18.9l-4.95-4.95a7 7 0 010-9.9zM10 11a2 2 0 100-4 2 2 0 000 4z"
-                        clipRule="evenodd"
-                      />
-                    </svg>
-                    <span className="text-sm">{property.location}</span>
-                  </div>
-
-                  <div className="flex items-center mb-3">
-                    <div className="flex items-center">
-                      <svg
-                        className="w-4 h-4 text-yellow-400"
-                        fill="currentColor"
-                        viewBox="0 0 20 20"
-                      >
-                        <path d="M9.049 2.927c.3-.921 1.603-.921 1.902 0l1.07 3.292a1 1 0 00.95.69h3.462c.969 0 1.371 1.24.588 1.81l-2.8 2.034a1 1 0 00-.364 1.118l1.07 3.292c.3.921-.755 1.688-1.54 1.118l-2.8-2.034a1 1 0 00-1.175 0l-2.8 2.034c-.784.57-1.838-.197-1.539-1.118l1.07-3.292a1 1 0 00-.364-1.118L2.98 8.72c-.783-.57-.38-1.81.588-1.81h3.461a1 1 0 00.951-.69l1.07-3.292z" />
-                      </svg>
-                      <span className="text-sm font-medium ml-1">
-                        {property.rating}
-                      </span>
-                    </div>
-                    <span className="text-sm text-gray-500 ml-2">
-                      ({property.reviews} reviews)
-                    </span>
-                  </div>
-
-                  <div className="flex flex-wrap gap-2 mb-4">
-                    {property.amenities.slice(0, 3).map((amenity, index) => (
-                      <span
-                        key={index}
-                        className="bg-gray-100 text-gray-800 px-2 py-1 rounded-md text-xs font-medium"
-                      >
-                        {amenity}
-                      </span>
-                    ))}
-                    {property.amenities.length > 3 && (
-                      <span className="bg-gray-100 text-gray-800 px-2 py-1 rounded-md text-xs font-medium">
-                        +{property.amenities.length - 3} more
-                      </span>
-                    )}
-                  </div>
-
-                  <button className="w-full bg-primary hover:bg-primary/90 text-white py-2 px-4 rounded-md font-medium transition-colors">
-                    View Details
-                  </button>
-                </div>
-              </div>
-            ))}
-          </div>
-        </div>
-      </section>
+      <FeaturedProperties />
 
       {/* How It Works */}
-      <section className="py-16 bg-white">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-          <div className="text-center mb-12">
-            <h2 className="text-3xl font-bold text-gray-900 mb-4">
-              How It Works
-            </h2>
-            <p className="text-gray-600 text-lg">
-              Simple steps to find your perfect PG
-            </p>
-          </div>
-
-          <div className="grid md:grid-cols-3 gap-8">
-            <div className="text-center group">
-              <div className="w-16 h-16 bg-primary rounded-full flex items-center justify-center mx-auto mb-6 group-hover:bg-primary/90 transition-colors">
-                <svg
-                  className="w-8 h-8 text-primary-foreground"
-                  fill="none"
-                  stroke="currentColor"
-                  viewBox="0 0 24 24"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth={2}
-                    d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z"
-                  />
-                </svg>
-              </div>
-              <h3 className="text-xl font-semibold text-gray-900 mb-3">
-                Search & Filter
-              </h3>
-              <p className="text-gray-600">
-                Search for PG accommodations by location, budget, and amenities.
-                Use our smart filters to find exactly what you need.
-              </p>
-            </div>
-
-            <div className="text-center group">
-              <div className="w-16 h-16 bg-primary rounded-full flex items-center justify-center mx-auto mb-6 group-hover:bg-primary/90 transition-colors">
-                <svg
-                  className="w-8 h-8 text-primary-foreground"
-                  fill="currentColor"
-                  viewBox="0 0 20 20"
-                >
-                  <path
-                    fillRule="evenodd"
-                    d="M10 18a8 8 0 100-16 8 8 0 000 16zm3.707-9.293a1 1 0 00-1.414-1.414L9 10.586 7.707 9.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l4-4z"
-                    clipRule="evenodd"
-                  />
-                </svg>
-              </div>
-              <h3 className="text-xl font-semibold text-gray-900 mb-3">
-                Compare & Choose
-              </h3>
-              <p className="text-gray-600">
-                Compare verified properties with genuine photos, reviews, and
-                transparent pricing. No hidden charges.
-              </p>
-            </div>
-
-            <div className="text-center group">
-              <div className="w-16 h-16 bg-primary rounded-full flex items-center justify-center mx-auto mb-6 group-hover:bg-primary/90 transition-colors">
-                <svg
-                  className="w-8 h-8 text-primary-foreground"
-                  fill="currentColor"
-                  viewBox="0 0 20 20"
-                >
-                  <path d="M10.707 2.293a1 1 0 00-1.414 0l-7 7a1 1 0 001.414 1.414L4 10.414V17a1 1 0 001 1h2a1 1 0 001-1v-2a1 1 0 011-1h2a1 1 0 011 1v2a1 1 0 001 1h2a1 1 0 001-1v-6.586l.293.293a1 1 0 001.414-1.414l-7-7z" />
-                </svg>
-              </div>
-              <h3 className="text-xl font-semibold text-gray-900 mb-3">
-                Book & Move In
-              </h3>
-              <p className="text-gray-600">
-                Book instantly with zero brokerage. Schedule a visit or move in
-                directly. We handle all the paperwork.
-              </p>
-            </div>
-          </div>
-        </div>
-      </section>
+      <HowItWorks />
 
       {/* Why Choose Us */}
-      <section className="py-16 bg-blue-50">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-          <div className="text-center mb-12">
-            <h2 className="text-3xl font-bold text-gray-900 mb-4">
-              Why Choose Us?
-            </h2>
-            <p className="text-gray-600 text-lg">
-              Your trusted partner for PG accommodation
-            </p>
-          </div>
-
-          <div className="grid md:grid-cols-2 lg:grid-cols-4 gap-8">
-            <div className="text-center">
-              <div className="w-12 h-12 bg-primary rounded-lg flex items-center justify-center mx-auto mb-4">
-                <svg
-                  className="w-6 h-6 text-white"
-                  fill="currentColor"
-                  viewBox="0 0 20 20"
-                >
-                  <path
-                    fillRule="evenodd"
-                    d="M2.166 4.999A11.954 11.954 0 0010 1.944 11.954 11.954 0 0017.834 5c.11.65.166 1.32.166 2.001 0 5.225-3.34 9.67-8 11.317C5.34 16.67 2 12.225 2 7c0-.682.057-1.35.166-2.001zm11.541 3.708a1 1 0 00-1.414-1.414L9 10.586 7.707 9.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l4-4z"
-                    clipRule="evenodd"
-                  />
-                </svg>
-              </div>
-              <h3 className="font-semibold text-gray-900 mb-2">
-                100% Verified
-              </h3>
-              <p className="text-gray-600 text-sm">
-                All properties are physically verified by our team
-              </p>
-            </div>
-
-            <div className="text-center">
-              <div className="w-12 h-12 bg-primary rounded-lg flex items-center justify-center mx-auto mb-4">
-                <svg
-                  className="w-6 h-6 text-white"
-                  fill="currentColor"
-                  viewBox="0 0 20 20"
-                >
-                  <path d="M13 6a3 3 0 11-6 0 3 3 0 016 0zM18 8a2 2 0 11-4 0 2 2 0 014 0zM14 15a4 4 0 00-8 0v3h8v-3z" />
-                </svg>
-              </div>
-              <h3 className="font-semibold text-gray-900 mb-2">
-                Zero Brokerage
-              </h3>
-              <p className="text-gray-600 text-sm">
-                No hidden charges or brokerage fees
-              </p>
-            </div>
-
-            <div className="text-center">
-              <div className="w-12 h-12 bg-primary rounded-lg flex items-center justify-center mx-auto mb-4">
-                <svg
-                  className="w-6 h-6 text-white"
-                  fill="none"
-                  stroke="currentColor"
-                  viewBox="0 0 24 24"
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth={2}
-                    d="M12 8v4l3 3m6-3a9 9 0 11-18 0 9 9 0 0118 0z"
-                  />
-                </svg>
-              </div>
-              <h3 className="font-semibold text-gray-900 mb-2">24/7 Support</h3>
-              <p className="text-gray-600 text-sm">
-                Round the clock customer support
-              </p>
-            </div>
-
-            <div className="text-center">
-              <div className="w-12 h-12 bg-primary rounded-lg flex items-center justify-center mx-auto mb-4">
-                <svg
-                  className="w-6 h-6 text-white"
-                  fill="currentColor"
-                  viewBox="0 0 20 20"
-                >
-                  <path
-                    fillRule="evenodd"
-                    d="M10 18a8 8 0 100-16 8 8 0 000 16zm3.707-9.293a1 1 0 00-1.414-1.414L9 10.586 7.707 9.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l4-4z"
-                    clipRule="evenodd"
-                  />
-                </svg>
-              </div>
-              <h3 className="font-semibold text-gray-900 mb-2">
-                Instant Booking
-              </h3>
-              <p className="text-gray-600 text-sm">
-                Book your PG instantly without any hassle
-              </p>
-            </div>
-          </div>
-        </div>
-      </section>
+      <WhyChooseUs />
 
       {/* Testimonials */}
-      <section className="py-16 bg-white">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-          <div className="text-center mb-12">
-            <h2 className="text-3xl font-bold text-gray-900 mb-4">
-              What Our Users Say
-            </h2>
-            <p className="text-gray-600 text-lg">
-              Real experiences from our happy customers
-            </p>
-          </div>
-
-          <div className="grid md:grid-cols-3 gap-8">
-            {testimonials.map((testimonial, index) => (
-              <div
-                key={index}
-                className="bg-white rounded-lg shadow-md border-0 hover:shadow-lg transition-shadow"
-              >
-                <div className="p-6">
-                  <div className="flex items-center mb-4">
-                    <Avatar className="w-12 h-12 rounded-full object-cover mr-4">
-                      <AvatarImage
-                        src={testimonial.avatar || "/placeholder.svg"}
-                      />
-                      <AvatarFallback>
-                        {testimonial.name.charAt(0)}
-                      </AvatarFallback>
-                    </Avatar>
-
-                    <div>
-                      <h4 className="font-semibold text-gray-900">
-                        {testimonial.name}
-                      </h4>
-                      <p className="text-sm text-gray-500">
-                        {testimonial.location}
-                      </p>
-                    </div>
-                  </div>
-                  <div className="flex mb-4">
-                    {[...Array(testimonial.rating)].map((_, i) => (
-                      <svg
-                        key={i}
-                        className="w-4 h-4 text-yellow-400"
-                        fill="currentColor"
-                        viewBox="0 0 20 20"
-                      >
-                        <path d="M9.049 2.927c.3-.921 1.603-.921 1.902 0l1.07 3.292a1 1 0 00.95.69h3.462c.969 0 1.371 1.24.588 1.81l-2.8 2.034a1 1 0 00-.364 1.118l1.07 3.292c.3.921-.755 1.688-1.54 1.118l-2.8-2.034a1 1 0 00-1.175 0l-2.8 2.034c-.784.57-1.838-.197-1.539-1.118l1.07-3.292a1 1 0 00-.364-1.118L2.98 8.72c-.783-.57-.38-1.81.588-1.81h3.461a1 1 0 00.951-.69l1.07-3.292z" />
-                      </svg>
-                    ))}
-                  </div>
-                  <p className="text-gray-600 italic">"{testimonial.text}"</p>
-                </div>
-              </div>
-            ))}
-          </div>
-        </div>
-      </section>
+      <Testimonials />
 
       {/* CTA Section */}
-      <section className="py-16 bg-primary">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 text-center">
-          <h2 className="text-3xl font-bold text-white mb-4">
-            Ready to Find Your Perfect PG?
-          </h2>
-          <p className="text-blue-100 text-lg mb-8 max-w-2xl mx-auto">
-            Join thousands of students and professionals who found their ideal
-            accommodation through our platform
-          </p>
-          <div className="flex flex-col sm:flex-row gap-4 justify-center">
-            <button className="bg-white text-primary hover:bg-gray-100 px-6 py-3 rounded-md font-semibold transition-colors">
-              Start Searching
-            </button>
-            <button className="border border-white text-white hover:bg-white hover:text-primary bg-transparent px-6 py-3 rounded-md font-semibold transition-colors">
-              List Your Property
-            </button>
-          </div>
-        </div>
-      </section>
-
-      {/* Footer */}
-      <footer className="bg-gray-900 text-white py-12">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-          <div className="grid md:grid-cols-3 gap-8">
-            <div>
-              <div className="flex items-center space-x-2 mb-4">
-                <div className="w-8 h-8 bg-primary rounded-lg flex items-center justify-center">
-                  <svg
-                    className="w-5 h-5 text-white"
-                    fill="currentColor"
-                    viewBox="0 0 20 20"
-                  >
-                    <path d="M10.707 2.293a1 1 0 00-1.414 0l-7 7a1 1 0 001.414 1.414L4 10.414V17a1 1 0 001 1h2a1 1 0 001-1v-2a1 1 0 011-1h2a1 1 0 011 1v2a1 1 0 001 1h2a1 1 0 001-1v-6.586l.293.293a1 1 0 001.414-1.414l-7-7z" />
-                  </svg>
-                </div>
-              </div>
-              <p className="text-gray-400 mb-4">
-                Your trusted platform for finding verified PG accommodations
-                across India.
-              </p>
-              <div className="flex space-x-4">
-                <div className="w-8 h-8 bg-gray-800 rounded-full flex items-center justify-center hover:bg-primary transition-colors cursor-pointer">
-                  <span className="text-sm">f</span>
-                </div>
-                <div className="w-8 h-8 bg-gray-800 rounded-full flex items-center justify-center hover:bg-primary transition-colors cursor-pointer">
-                  <span className="text-sm">t</span>
-                </div>
-                <div className="w-8 h-8 bg-gray-800 rounded-full flex items-center justify-center hover:bg-primary transition-colors cursor-pointer">
-                  <span className="text-sm">in</span>
-                </div>
-              </div>
-            </div>
-
-            <div>
-              <h3 className="font-semibold mb-4">Quick Links</h3>
-              <ul className="space-y-2 text-gray-400">
-                <li>
-                  <a href="#" className="hover:text-white transition-colors">
-                    Properties
-                  </a>
-                </li>
-                <li>
-                  <a href="#" className="hover:text-white transition-colors">
-                    How It Works
-                  </a>
-                </li>
-                <li>
-                  <a href="#" className="hover:text-white transition-colors">
-                    List Property
-                  </a>
-                </li>
-              </ul>
-            </div>
-
-            <div>
-              <h3 className="font-semibold mb-4">Support</h3>
-              <ul className="space-y-2 text-gray-400">
-                <li>
-                  <a href="#" className="hover:text-white transition-colors">
-                    Help Center
-                  </a>
-                </li>
-                <li>
-                  <a href="#" className="hover:text-white transition-colors">
-                    Terms of Service
-                  </a>
-                </li>
-                <li>
-                  <a href="#" className="hover:text-white transition-colors">
-                    Privacy Policy
-                  </a>
-                </li>
-                <li>
-                  <a href="#" className="hover:text-white transition-colors">
-                    FAQ
-                  </a>
-                </li>
-              </ul>
-            </div>
-          </div>
-
-          <div className="border-t border-gray-800 mt-8 pt-8 text-center text-gray-400">
-            <p>&copy; 2024 PG Accommodation Platform. All rights reserved.</p>
-          </div>
-        </div>
-      </footer>
+      <Cta />
     </div>
   );
 }
