+++ apps/web/components/footer.tsx
@@ -1,72 +1,89 @@
 "use client";
 
-import { Button } from "../../../packages/ui/src/components/button";
-import { Input } from "../../../packages/ui/src/components/input";
-import { Textarea } from "../../../packages/ui/src/components/textarea";
+import Brand from "./Brand";
 
 export default function Footer() {
   return (
-    <footer
-      className="w-full text-white py-8"
-      style={{ backgroundColor: "oklch(0.5113 0.212 275.57)" }}
-    >
-      <div className="w-full">
-        <div className="max-w-7xl mx-auto px-4">
-          <div className="flex flex-col md:flex-row justify-between items-start lg:items-center gap-8">
-            <div className="mb-6 md:mb-0 p-4">
-              <h2 className="text-2xl font-bold mb-2">Contact Us</h2>
-              <p className="mb-1">
-                Email:{" "}
-                <a href="mailto:baasthan@gmail.com" className="underline">
-                  baasthan@gmail.com
-                </a>
-              </p>
-              <p>
-                Phone 1:{" "}
-                <a href="tel:+916295626915" className="underline">
-                  +91 62956 26915
-                </a>
-              </p>
-              <p>
-                Phone 2:{" "}
-                <a href="tel:+916290654204" className="underline">
-                  +91 62906 54204
-                </a>
-              </p>
+    <footer className="bg-gray-900 text-white py-12">
+      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
+        <div className="grid md:grid-cols-3 gap-8">
+          <div>
+            <div className="flex items-center space-x-2 mb-4">
+              <div className="w-8 h-8 rounded-lg flex items-center justify-center">
+                <Brand />
+              </div>
+              <div className=" font-semibold text-2xl">
+                Baasthan Estate Pvt. Ltd.
+              </div>
             </div>
+            <p className="text-gray-400 mb-4">
+              Your trusted platform for finding verified PG accommodations
+              across India.
+            </p>
+            <div className="flex space-x-4">
+              <div className="w-8 h-8 bg-gray-800 rounded-full flex items-center justify-center hover:bg-primary transition-colors cursor-pointer">
+                <span className="text-sm">f</span>
+              </div>
+              <div className="w-8 h-8 bg-gray-800 rounded-full flex items-center justify-center hover:bg-primary transition-colors cursor-pointer">
+                <span className="text-sm">t</span>
+              </div>
+              <div className="w-8 h-8 bg-gray-800 rounded-full flex items-center justify-center hover:bg-primary transition-colors cursor-pointer">
+                <span className="text-sm">in</span>
+              </div>
+            </div>
+          </div>
 
-            <form className="flex flex-col gap-4 w-full max-w-sm  p-4 rounded ">
-              <h3 className="text-lg font-semibold mb-2">Send us a message</h3>
-              <Input
-                name="name"
-                type="text"
-                placeholder="Your Name"
-                className=" placeholder:text-white"
-                required
-              />
-              <Input
-                name="email"
-                type="email"
-                placeholder="Your Email"
-                className=" placeholder:text-white"
-                required
-              />
-              <Textarea
-                name="message"
-                placeholder="Your Message"
-                className=" placeholder:text-white"
-                rows={3}
-                required
-              />
-              {/* TODO: Add a variant that works well with primary color */}
-              <Button type="submit" variant={"secondary"}>
-                Send
-              </Button>
-            </form>
+          <div>
+            <h3 className="font-semibold mb-4">Quick Links</h3>
+            <ul className="space-y-2 text-gray-400">
+              <li>
+                <a href="#" className="hover:text-white transition-colors">
+                  Properties
+                </a>
+              </li>
+              <li>
+                <a href="#" className="hover:text-white transition-colors">
+                  How It Works
+                </a>
+              </li>
+              <li>
+                <a href="#" className="hover:text-white transition-colors">
+                  List Property
+                </a>
+              </li>
+            </ul>
           </div>
-          <div className="text-center text-gray-200 mt-6 text-sm">
-            &copy; {new Date().getFullYear()} Baasthan. All rights reserved.
+
+          <div>
+            <h3 className="font-semibold mb-4">Support</h3>
+            <ul className="space-y-2 text-gray-400">
+              <li>
+                <a href="#" className="hover:text-white transition-colors">
+                  Help Center
+                </a>
+              </li>
+              <li>
+                <a href="#" className="hover:text-white transition-colors">
+                  Terms of Service
+                </a>
+              </li>
+              <li>
+                <a href="#" className="hover:text-white transition-colors">
+                  Privacy Policy
+                </a>
+              </li>
+              <li>
+                <a href="#" className="hover:text-white transition-colors">
+                  FAQ
+                </a>
+              </li>
+            </ul>
           </div>
+        </div>
+
+        <div className="border-t border-gray-800 mt-8 pt-8 text-center text-gray-400">
+          <p>&copy; 2025 Baasthan Estate Pvt. Ltd.</p>
+          <p> All rights reserved.</p>
         </div>
       </div>
     </footer>
